feat: use custom DNS resolver in domain accessibility verification http client
